Arreglado el controller y el service de application

+++ Back/src/main/java/com/campusdual/bfp/model/dao/ApplicationDao.java
@@ -3,5 +3,5 @@ package com.campusdual.bfp.model.dao;
 import com.campusdual.bfp.model.Application;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ApplicationDao extends JpaRepository<Application, Integer> {
+public interface ApplicationDao extends JpaRepository<Application, Long> {
 }
